requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
anthropic==0.7.0
openai==1.3.0
google-generativeai==0.3.0
cohere==4.37
httpx==0.25.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3
aiofiles==23.2.1
